personalized-voice-chatbot/requirements.txt
Flask==2.3.2
faster-whisper==0.10.0
coqui-tts==0.22.0
torch>=1.10
numpy
soundfile
pydub
python-dotenv
